update:修改密码校验

+++ lib/pages/personal_center/modify_passwordbycode.dart
@@ -3,10 +3,12 @@ import 'package:bid/service/service_method.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:bid/common/toast.dart';
+import 'package:bid/common/count_down.dart';
 import 'package:flutter/physics.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../common/log_utils.dart';
 import 'package:sprintf/sprintf.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class ModifyPasswordByCode extends StatefulWidget {
   @override
@@ -14,6 +16,9 @@ class ModifyPasswordByCode extends StatefulWidget {
 }
 
 class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
+  bool changeCount = false;
+  bool autoValidate = false;
+  bool mobileValidate = false;
   final setPawwordFormKey = GlobalKey<FormState>();
   var params = {
     'mobile': '',
@@ -35,8 +40,47 @@ class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
     );
   }
 
+  // 密码校验
+  String validatePwd(String value) {
+    print('密码验证码$value');
+    // 正则匹配手机号
+    RegExp exp = RegExp(r'^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[^]{8,18}$');
+
+    if (value.isEmpty) {
+      return '密码不能为空';
+    } else {
+      if (!exp.hasMatch(value)) {
+        return '8-18位数字大小写字母的组合';
+      } else {
+        return null;
+      }
+    }
+  }
+
+  // 验证手机号码
+  String validateMibile(String value) {
+    errorMobileText = null;
+    print('手机校验$value==');
+    // 正则匹配手机号
+    RegExp exp = RegExp(
+        r'^((13[0-9])|(14[0-9])|(15[0-9])|(16[0-9])|(17[0-9])|(18[0-9])|(19[0-9]))\d{8}$');
+
+    if (value.isEmpty) {
+      print('手机校验为空====');
+      return '手机号码不能为空';
+    } else {
+      if (!exp.hasMatch(value)) {
+        print('手机校验规则不对');
+        return '请输入正确的账号';
+      } else {
+        return null;
+      }
+    }
+  }
+
   // 获取验证码
   void _getPhoneCode() async {
+    mobileValidate = true;
     setPawwordFormKey.currentState.save();
     if (params['mobile'].isEmpty) {
       setState(() {
@@ -47,15 +91,18 @@ class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
         errorMobileText = null;
       });
     }
-
-    var data = {'mobile': params['mobile']};
-    print('获取验证码注册$data');
-    await requestGet('getModifyPasswordCode', formData: data).then((val) {
-      Toast.toast(
-        context,
-        msg: val['message'],
-      );
-    });
+    var flag =
+        validateMibile(params['mobile'].toString()) == null ? true : false;
+    if (flag) {
+      var data = {'mobile': params['mobile']};
+      print('获取验证码注册$data');
+      await requestGet('getModifyPasswordCode', formData: data).then((val) {
+        Toast.toast(
+          context,
+          msg: val['message'],
+        );
+      });
+    }
   }
 
   Widget _buildModifyPwdTextForm() {
@@ -78,6 +125,166 @@ class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
     );
   }
 
+  Widget _buildRow(item) {
+    return new Container(
+      decoration: BoxDecoration(
+        color: Colors.white,
+        // border: Border(
+        //   bottom: BorderSide(
+        //     width: 1,
+        //     color: Color(0xffe5e5e5),
+        //   ),
+        // ),
+      ),
+      child: new Row(
+        children: [
+          new Expanded(
+            child: new Container(
+              padding: EdgeInsets.all(15),
+              // height: 30.0,
+              decoration: BoxDecoration(
+                  //border: new Border.all(color: Colors.red),
+                  ),
+              // child: new Expanded(
+              child: TextFormField(
+                controller: TextEditingController.fromValue(
+                  TextEditingValue(
+                    text:
+                        '${params[item['value']] == null ? "" : params[item['value']]}',
+                    // 保持光标在最后
+                    selection: TextSelection.fromPosition(
+                      TextPosition(
+                          affinity: TextAffinity.downstream,
+                          offset: '${params[item['value']]}'.length),
+                    ),
+                  ),
+                ),
+                decoration: InputDecoration(
+                  errorText: item['value'] == 'mobile' ? errorMobileText : null,
+                  prefixIcon: Container(
+                    width: ScreenUtil().setWidth(140.0),
+                    // padding: EdgeInsets.only(right: 20),
+                    child: Center(
+                      child: RichText(
+                        textAlign: TextAlign.left,
+                        text: TextSpan(
+                          text: '*',
+                          style:
+                              TextStyle(color: Color.fromRGBO(255, 113, 66, 1)),
+                          children: <TextSpan>[
+                            TextSpan(
+                              text: '${item['label']}',
+                              style: TextStyle(
+                                decoration: TextDecoration.none,
+                                color: Color(0xFF888888),
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold,
+                                //fontFamily: defaultFontFamily,
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                  ),
+                  suffixIcon: _showText(item),
+                  // border: InputBorder.none,
+                ),
+                //controller: controller,
+                //maxLength: 30, //最大长度，设置此项会让TextField右下角有一个输入数量的统计字符串
+                maxLines: null, //最大行数
+                autocorrect: true, //是否自动更正
+                autofocus: false, //是否自动对焦
+                autovalidate:
+                    item['value'] == 'mobile' ? mobileValidate : autoValidate,
+                obscureText: false, //是否是密码
+                textAlign: TextAlign.left, //文本对齐方式
+                style: TextStyle(
+                  fontSize: 20.0,
+                ), //输入文本的样式
+                //inputFormatters: [WhitelistingTextInputFormatter.digitsOnly],//允许的输入格式
+                onChanged: (v) {
+                  if (item['value'] == 'mobile') {
+                    params[item['value']] = v;
+                    if (v.length == 11) {
+                      setPawwordFormKey.currentState.save();
+                      setState(() {
+                        changeCount =
+                            validateMibile(params[item['value']].toString()) ==
+                                    null
+                                ? true
+                                : false;
+                      });
+                    }
+                  } else if (item['value'] == 'newPwd') {
+                    setPawwordFormKey.currentState.save();
+                  }
+                },
+                onSaved: (val) {
+                  var key = item['value'];
+                  setState(() {
+                    params[key] = val;
+                  });
+                },
+                validator: (value) {
+                  if (item['value'] == 'mobile') {
+                    // 正则匹配手机号
+                    RegExp exp = RegExp(
+                        r'^((13[0-9])|(14[0-9])|(15[0-9])|(16[0-9])|(17[0-9])|(18[0-9])|(19[0-9]))\d{8}$');
+                    if (value.isEmpty) {
+                      return '手机号码不能为空';
+                    } else {
+                      if (!exp.hasMatch(value)) {
+                        return '请输入正确的账号';
+                      } else {
+                        return null;
+                      }
+                    }
+                  } else if (item['value'] == 'newPwd') {
+                    print('密码验证码$value');
+                    // 正则匹配手机号
+                    RegExp exp =
+                        RegExp(r'^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[^]{8,18}$');
+
+                    if (value.isEmpty) {
+                      return '密码不能为空';
+                    } else {
+                      if (!exp.hasMatch(value)) {
+                        return '8-18位数字大小写字母的组合';
+                      } else {
+                        return null;
+                      }
+                    }
+                  } else if (item['value'] == 'confirmNewpwd') {
+                    // setPawwordFormKey.currentState.save();
+                    if (value.isEmpty) {
+                      return "不能为空";
+                    } else {
+                      if (value != params['newPwd']) {
+                        return '两次密码不一致';
+                      } else {
+                        return null;
+                      }
+                    }
+                  } else {
+                    if (value.isEmpty) {
+                      return "不能为空";
+                    }
+                    return null;
+                  }
+                },
+                enabled: true, //是否禁用
+              ),
+            ),
+            // ),
+          ),
+          // _showText(item),
+        ],
+      ),
+    );
+  }
+
+/*
   Widget _buildRow(item) {
     return new Container(
       decoration: BoxDecoration(
@@ -147,38 +354,48 @@ class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
               ),
             ),
           ),
-          _showText(item),
+         
         ],
       ),
     );
   }
-
+*/
   Widget _showText(item) {
     if (item['label'] == "验证码:") {
       return Container(
         padding: EdgeInsets.only(right: 20),
-        child: FlatButton(
-          // color: Colors.blue,
-          highlightColor: Colors.blue[700],
-          // colorBrightness: Brightness.dark,
-          splashColor: Colors.grey,
-          child: Text(
-            '获取验证码',
-            style: TextStyle(
-              color: Color(0xFF4389ED),
-            ),
-          ),
-          shape: RoundedRectangleBorder(
-            side: BorderSide(
-              color: Color(0xFF4389ED),
-              width: 1,
-            ),
-            borderRadius: BorderRadius.circular(20.0),
-          ),
-          onPressed: () {
+        child: LoginFormCode(
+          key: childKey,
+          isChange: changeCount,
+          countdown: 60,
+          available: true,
+          onTapCallback: (val) {
+            print('父组件拿到子组件的方法$val');
             _getPhoneCode();
           },
         ),
+        // child: FlatButton(
+        //   // color: Colors.blue,
+        //   highlightColor: Colors.blue[700],
+        //   // colorBrightness: Brightness.dark,
+        //   splashColor: Colors.grey,
+        //   child: Text(
+        //     '获取验证码',
+        //     style: TextStyle(
+        //       color: Color(0xFF4389ED),
+        //     ),
+        //   ),
+        //   shape: RoundedRectangleBorder(
+        //     side: BorderSide(
+        //       color: Color(0xFF4389ED),
+        //       width: 1,
+        //     ),
+        //     borderRadius: BorderRadius.circular(20.0),
+        //   ),
+        //   onPressed: () {
+        //     _getPhoneCode();
+        //   },
+        // ),
       );
     } else {
       return Container(
@@ -235,44 +452,48 @@ class _ModifyPasswordByCodeState extends State<ModifyPasswordByCode> {
   }
 
   void _submit() async {
+    autoValidate = true;
+    mobileValidate = true;
     setPawwordFormKey.currentState.save();
-    var data = {
-      "newPwd": params['newPwd'], //新密码
-      "checkCode": params['checkCode'], //验证码
-      "mobile": params['mobile']
-    };
+    if ((setPawwordFormKey.currentState as FormState).validate()) {
+      var data = {
+        "newPwd": params['newPwd'], //新密码
+        "checkCode": params['checkCode'], //验证码
+        "mobile": params['mobile']
+      };
 
-    print('修改密码$data');
-    await request('resetPwdByCode', formData: data).then((val) {
-      print('----------------------$val');
-      if (val['code'] == 0) {
-        showDialog(
-            context: context,
-            builder: (context) {
-              return AlertDialog(
-                // title: Text('退出登录提示'),
-                content: Text('您已成功修改密码，请重新登录!'),
-                actions: <Widget>[
-                  FlatButton(
-                    child: Text('确认'),
-                    onPressed: () async {
-                      final prefs = await SharedPreferences.getInstance();
-                      final result = await prefs.clear();
-                      if (result) {
-                        Navigator.of(context).pop('cancel');
-                        Application.router.navigateTo(context, '/sigin');
-                      }
-                    },
-                  ),
-                ],
-              );
-            });
-      } else {
-        Toast.toast(
-          context,
-          msg: val['message'],
-        );
-      }
-    });
+      print('修改密码$data');
+      await request('resetPwdByCode', formData: data).then((val) {
+        print('----------------------$val');
+        if (val['code'] == 0) {
+          showDialog(
+              context: context,
+              builder: (context) {
+                return AlertDialog(
+                  // title: Text('退出登录提示'),
+                  content: Text('您已成功修改密码，请重新登录!'),
+                  actions: <Widget>[
+                    FlatButton(
+                      child: Text('确认'),
+                      onPressed: () async {
+                        final prefs = await SharedPreferences.getInstance();
+                        final result = await prefs.clear();
+                        if (result) {
+                          Navigator.of(context).pop('cancel');
+                          Application.router.navigateTo(context, '/sigin');
+                        }
+                      },
+                    ),
+                  ],
+                );
+              });
+        } else {
+          Toast.toast(
+            context,
+            msg: val['message'],
+          );
+        }
+      });
+    }
   }
 }
